requirements.txt
googletrans==4.0.0rc1
numpy==1.26.4
Pillow==10.2.0
streamlit==1.32.1
opencv-python-headless
easyocr